Split off clifford.cc from clifford.h and src/arch/cc_light/*compiler.h
Furthermore added clifford_optimize calls around prescheduler.
And added options to control that.
In test_mapper set such and option
